+++ core_client/__init__.py
@@ -6,7 +6,11 @@ import base64
 import json
 from datetime import datetime
 from httpx import InvalidURL as HttpInvalidURL, HTTPError
-from pydantic import HttpUrl, ValidationError as PydanticValidationError, validate_arguments
+from pydantic import (
+    HttpUrl,
+    ValidationError as PydanticValidationError,
+    validate_arguments,
+)
 
 from . import base
 from .models import Client as ClientModel
@@ -22,6 +26,7 @@ class Client:
         password: str = None,
         access_token: str = None,
         refresh_token: str = None,
+        auth0_token: str = None,
         retries: int = 3,
         timeout: float = 10.0,
     ):
@@ -37,45 +42,109 @@ class Client:
         self.password = password
         self.access_token = access_token
         self.refresh_token = refresh_token
+        self.access_token_expires_at = None
+        self.auth0_token = auth0_token
         self.retries = retries
         self.timeout = timeout
 
-    def _access_token(self, response: Token = None):
-        if response:
-            self.token_expire = json.loads(
-                (base64.b64decode(response.access_token.split(".")[1])).decode("utf-8")
-            )["exp"]
-            self.access_token = response.access_token
+    def _basic_login(self):
+        r_login = httpx.post(
+            url=f"{self.base_url}/api/login",
+            json={
+                "username": f"{self.username}",
+                "password": f"{self.password}",
+            },
+            # failed login delay: 5s
+            timeout=10.0,
+        )
+        if r_login.status_code == 200:
+            try:
+                response = Token(**r_login.json())
+                self._set_access_token_expires_at(response)
+                self.refresh_token = response.refresh_token
+                return response
+            except PydanticValidationError:
+                return Token()
+        else:
+            raise HTTPError("Authorization failed")
 
-    def _refresh_token(self):
+    def _auth0_login(self):
+        _headers = self.headers
+        _headers["authorization"] = f"Bearer {self.auth0_token}"
+        r_login = httpx.post(
+            url=f"{self.base_url}/api/login",
+            headers=_headers,
+            # failed login delay: 5s
+            timeout=10.0,
+        )
+        if r_login.status_code == 200:
+            try:
+                response = Token(**r_login.json())
+                self._set_access_token_expires_at(response)
+                self.refresh_token = response.refresh_token
+                return response
+            except PydanticValidationError:
+                return Token()
+        else:
+            raise HTTPError("Authorization failed")
+
+    def _set_access_token_expires_at(self, response: Token = None):
+        if response:
+            try:
+                self.access_token_expires_at = json.loads(
+                    (
+                        base64.b64decode(response.access_token.split(".")[1])
+                    ).decode("utf-8")
+                )["exp"]
+            except (AttributeError, IndexError):
+                raise HTTPError("Authorization failed")
+            self.access_token = response.access_token
+            return self.access_token_expires_at
+
+    def _access_token_is_expired(self):
+        if not self.access_token_expires_at:
+            self._set_access_token_expires_at(
+                Token(access_token=self.access_token)
+            )
+        if (
+            datetime.fromtimestamp(self.access_token_expires_at)
+            > datetime.now()
+        ):
+            return False
+        return True
+
+    def _refresh_access_token(self):
         if self.refresh_token:
             _headers = self.headers
             _headers["authorization"] = f"Bearer {self.refresh_token}"
-            r_refresh_token = httpx.get(
+            r_refresh_access_token = httpx.get(
                 url=f"{self.base_url}/api/login/refresh",
                 headers=_headers,
                 # failed login delay: 5s
                 timeout=10.0,
             )
-            if r_refresh_token.status_code == 200:
-                response = AccessToken(**r_refresh_token.json())
-                self._access_token(response)
-                return True
-        return False
-
-    def _token_expired(self):
-        if datetime.fromtimestamp(self.token_expire) > datetime.now():
-            return False
-        return True
+            if r_refresh_access_token.status_code == 200:
+                response = AccessToken(**r_refresh_access_token.json())
+                if response.access_token is not None:
+                    self._set_access_token_expires_at(response)
+            else:
+                self._basic_login()
+        else:
+            self._basic_login()
 
     def _get_headers(self):
         _headers = self.headers
-        if self.refresh_token:
-            if self._token_expired():
-                self._refresh_token()
-            _headers["authorization"] = f"Bearer {self.access_token}"
-            return _headers
+        if self.refresh_token and self._access_token_is_expired():
+            self._refresh_access_token()
+        _headers["authorization"] = f"Bearer {self.access_token}"
         return _headers
+
+    def token(self):
+        return Token(
+            access_token=self.access_token,
+            refresh_token=self.refresh_token,
+            expires_at=int(self.access_token_expires_at),
+        )
 
     def login(self):
         r_about = httpx.get(url=f"{self.base_url}/api")
@@ -84,25 +153,32 @@ class Client:
                 about = About(**r_about.json())
             except PydanticValidationError:
                 raise HttpInvalidURL(f'"{self.base_url}/api"')
-            if about.auths and "localjwt" in about.auths and self.username and self.password:
-                r_login = httpx.post(
-                    url=f"{self.base_url}/api/login",
-                    json={
-                        "username": f"{self.username}",
-                        "password": f"{self.password}",
-                    },
-                    # failed login delay: 5s
-                    timeout=10.0,
-                )
-                if r_login.status_code == 200:
+            if about.auths and "localjwt" in about.auths:
+                if self.refresh_token:
                     try:
-                        response = Token(**r_login.json())
-                        self._access_token(response)
-                        self.refresh_token = response.refresh_token
+                        self._refresh_access_token()
+                        return self.token()
+                    except (PydanticValidationError, TypeError):
+                        raise HTTPError("Authorization failed")
+                elif self.auth0_token:
+                    self._auth0_login()
+                    return self.token()
+                elif self.username and self.password:
+                    self._basic_login()
+                    return self.token()
+                elif (
+                    self.access_token
+                    and not self.refresh_token
+                    and not (self.username and self.password)
+                ):
+                    try:
+                        response = Token(access_token=self.access_token)
+                        self._set_access_token_expires_at(response)
                         return response
-                    except PydanticValidationError:
+                    except (PydanticValidationError, TypeError):
                         return Token()
-                raise HTTPError(f'"Authorization failed", {r_login.status_code}')
+                else:
+                    raise HTTPError("Authorization failed")
             return Token()
         raise HTTPError(f'"{self.base_url}/api", {r_about.status_code}')
 
@@ -150,6 +226,8 @@ for module_info in pkgutil.walk_packages(base.__path__):
         submodule_name = f"{module_name}.{submodule_info.name}"
         submodule = importlib.import_module(submodule_name)
         if hasattr(submodule, "asyncio"):
-            AsyncClient._add_proxy_method(submodule_info.name, submodule.asyncio)
+            AsyncClient._add_proxy_method(
+                submodule_info.name, submodule.asyncio
+            )
         if hasattr(submodule, "sync"):
             Client._add_proxy_method(submodule_info.name, submodule.sync)