+++ core_client/base/models/v3/config_api_auth_auth0_tenant.py
@@ -1,0 +1,19 @@
+from pydantic import BaseModel
+
+
+class ConfigApiAuthAuth0Tenant(BaseModel):
+    """
+    {
+        "audience": "string",
+        "clientid": "string",
+        "domain": "string",
+        "users": [
+            "string"
+        ]
+    }
+    """
+
+    audience: str
+    clientid: str
+    domain: str
+    users: list[str]