+++ core_client/base/models/v3/metrics_collection_list.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+from . import MetricsCollection
+
+
+class MetricsCollectionList(BaseModel):
+    __root__: list[MetricsCollection]