+++ tests/6_post.py
@@ -6,7 +6,9 @@ from core_client.base.models import Token
 from core_client.base.models.v3 import ConfigSaved
 
 core_url = os.getenv("CORE_URL", "http://127.0.0.1:8080")
-client = Client(base_url=f"{core_url}", username="admin", password="test", timeout=10.0)
+client = Client(
+    base_url=f"{core_url}", username="admin", password="test", timeout=20.0
+)
 
 
 def test_token():
@@ -27,6 +29,7 @@ none_jwt_config = {
     "srt": {"enable": False},
     "rtmp": {"enable": False},
     "sessions": {"ip_ignorelist": []},
+    "version": 3,
 }